Prevent displaying decimal pont (ugly): convert to int when get as string

+++ src/main/java/com/codecool/shop/model/LineItem.java
@@ -26,7 +26,7 @@ public class LineItem {
     }
 
     public String getTotalPriceString() {
-        return getTotalPrice() + " " + product.getDefaultCurrency().toString();
+        return (int) getTotalPrice() + " " + product.getDefaultCurrency().toString();
     }
 
     public Product getProduct() {
